backend/requirements.txt
# 后端依赖
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiohttp>=3.9.0
pydantic>=2.0.0

# 二维码扫描
qreader>=3.0.0
opencv-python>=4.5.0
pdf2image>=1.16.0
Pillow>=8.0.0

# PDF处理
PyPDF2>=3.0.0

# 网页爬取
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.0.0
webdriver-manager>=4.0.0

# 其他
openpyxl>=3.0.0